+++ database/migrations/2020_11_06_192734_create_posts_table.php
@@ -23,6 +23,7 @@ class CreatePostsTable extends Migration {
             $table->boolean("is_thread")->default(false);
             $table->integer("replies")->default(0);
             $table->boolean("is_sticky")->default(false);
+            $table->dateTime("latest_reply")->nullable();
             $table->timestamps();
         });
     }